backend/pyproject.toml
[tool.poetry]
package-mode = false
name = "hackathon-pw-backend"
version = "0.1.0"
description = "Algorand ASA / NFT backend"
authors = ["team"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pyteal = "0.24.1"
beaker-pyteal = "1.1.1"
algokit-utils = "2.3.0"
setuptools = "<81"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = {extras = ["email"], version = "^2.5.0"}
py-algorand-sdk = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

